Restore support for UDP and multi-threading
In the case of UDP, a `msg`'s `address` appears to be `free()`d somewhere after
the `msg` has been dispatched, leading to problems if attempting to access
the `url`, `hostname`, or `port` of the `address` from a different thread.

In this case restoring the original logic path, but only in the case of UDP,
works.