Expand one-liner python version check in _callback_num_args
Apparently some combination of python 3.11 and cython [0.29?] doesn't handle
the one-liner correctly, causing the `X` in `X if Y else Z` to be resolved even
when Y was False.

I (s0600204) never had this problem, hence the "apparently".

It could be that the users reporting the problem were using the release version
of 0.10.0, which doesn't have the one-liner, and thus always uses the defunct
`_inspect.getargspec`, causing the issue.

See also:
* https://github.com/dsacre/pyliblo/issues/29
* https://github.com/Houston4444/RaySession/issues/152#issuecomment-1356896533